feat: basic qwen edit trainer works now
